+++ History.md
@@ -3,6 +3,7 @@
 * Moved trait parsing logic to the global level
 * Added [Improvely](http://www.improvely.com/)
 * Added [Get Satisfaction](https://getsatisfaction.com/)
+* Added [Visual Website Optimizer](http://visualwebsiteoptimizer.com/)
 
 0.10.5 / 2013-5-22
 ==================